Refactor Memory management in MemoryLayout, MemoryPool, and MemorySlot; enhance slot handling with new methods for better encapsulation and clarity, and improve performance by optimizing memory access patterns

+++ Thanos/Memory/MemoryLayout.cs
@@ -54,7 +54,7 @@ public readonly struct MemoryLayout
             Bitboards = (int)sizeOfBitboards;
             // Lo slot ora contiene solo i dati reali
             Slot = Node + Bitboards + Snakes;
-            Pool = (nuint)((long)Slot * maxNodes);
+            Pool = (nuint)(Slot * maxNodes);
         }
     }
